add an argument for BaaS.login
add an argument for logining without geting user infomation;
move login interface from bassRequest module to user module
